+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from "vscode";
 import registerLinter from "./linter";
 import registerFormatter from "./formatter";
 import registerCompletions from "./completion";
+import registerHover from "./hover";
 
 /**
  * Activate the extension. Run automatically by VSCode based on
@@ -10,6 +11,7 @@ import registerCompletions from "./completion";
 export function activate(context: vscode.ExtensionContext): void {
   registerFormatter(context);
   registerLinter(context);
+  registerHover(context);
   registerCompletions(context);
 }